refactor(document processor): Move get_page_count to lib
