+++ cl/corpus_importer/tasks.py
@@ -57,7 +57,8 @@ from cl.recap.mergers import find_docket_object, update_docket_metadata, \
     add_docket_entries, add_parties_and_attorneys, process_orphan_documents, \
     add_claims_to_docket, add_bankruptcy_data_to_docket
 from cl.scrapers.models import PACERFreeDocumentLog, PACERFreeDocumentRow
-from cl.scrapers.tasks import extract_recap_pdf, get_page_count
+from cl.scrapers.tasks import extract_recap_pdf
+from cl.lib.document_processors import get_page_count
 from cl.search.models import DocketEntry, RECAPDocument, Court, Docket, Tag, \
     ClaimHistory
 from cl.search.tasks import add_items_to_solr
