+++ cl/search/management/commands/cl_count_recap_pages.py
@@ -2,7 +2,7 @@ from collections import Counter
 
 
 from cl.lib.command_utils import VerboseCommand, logger
-from cl.scrapers.tasks import get_page_count
+from cl.lib.document_processors import get_page_count
 from cl.search.models import RECAPDocument
